WorkfolioWeb/backend/requirements.txt
flask
flask_cors
opencv-python
face_recognition
openpyxl
numpy
Pillow
